requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
pandas==2.1.4
openpyxl==3.1.2
pyserial==3.5
python-socketio==5.11.0
python-engineio>=4.8.0
eventlet==0.35.2
pyinstaller==6.3.0